Extend `Connector` lifetime for the whole `onReceive`
calling `ISocket::switchPlug` replaces the plug pointer
on the socket manager side with a new one and returns
the current owning pointer to the caller. It is the
caller responsibility to keep the current owning pointer
on itself after calling `ISocket::switchPlug` to avoid
using `this` pointing to a destructed object.

Bug: 353756694
Test: devices_unittests
Change-Id: I3da44c186cf181b0fb6d2b0507cc7bd422857b9c
Signed-off-by: Roman Kiryanov <rkir@google.com>