Update documentation
This updates the documentation to include a section on how to create
a JSON compilation database.

This can be useful for developers that would like to make use of fancy
features it provides.

Includes a fix for broken help `-h`, and defaults that point to the
root of the WORKSPACE vs. the current directory when generating the
compile_commands.json

Change-Id: I676c70b382508b0363e196b99f8c0206e25cabff