Log: Use abseil logging backend
This switches us over to log everything to the abseil logging backend,
this will enable us to install arbitrary sinks that can be used to
inform studio of user messages.

Currently we have one sink that will log colors on TTY and standard
logging otherwise. Note that we do not use the abseil logging format, as
studio is parsing our logs.

This sets us up for a unified log system that our dependencies
(Tink/gRPC/...) are converging on.

Bug: 358381877
Change-Id: I165606ab952f6fbeb4051e290493933c86560e45