vsock: Enable adb through vsock proxy
This introduces a virtio-goldfish-hostfwd-socket that can be used to
proxy a host port to a guest port. This includes a port of the qemu
looper which will allow us to run the async sockets on the qemu threads.

This allows us to proxy port from the host to port 5555 inside the
guest, where the adbd lives.

Test: Can connect via adb shell.
Bug: 331947598
Change-Id: I048b41b10e7bbd37399506e38957defdb74abeff