Fix incorrect file name for vlog in abseil c bridge
The VLOG macro introduces global static variables for each file that is
compiled. Unfortunately we will not be able to this for the c-version,
so instead we will dynamically allocate the vlogsite.

This will make sure that we can actually filter on file name of the file
using the c version.

Change-Id: Ie46a01281120bfb28fc933f933d710abd5656c00