Improve AbseilLogBridge performance x11.9
We now apply the same trick used by abseil itself, where we now store
the VLogSite pointer in a static variable along with the logging code,
this removes the need for hashmaps, and makes sure we allocate an object
only once.

Furthermore we solve a concurrent access issue where multiple threads
could corrupt the hashmap.

Includes a performance test to get the performance numbers, which
resulted in the following:

// Apple M1 Pro 2021
// BM_VLOG          7.48 ns         7.48 ns     93369436

// Old: BM_ALOGV          109 ns          109 ns      6441698
// New: BM_ALOGV         9.13 ns         9.10 ns     72448019
// 109 / 9.10 = 11.978021978021978
// 109 / 9.13 = 11.938663745892662
// C++ vs C: 7.48 / 9.13 = 0.81

Change-Id: I7a09d0a4c72b1c256ac97703673150246543a37f