Improve vsock performance
RW packets can arrive fragmented, as an array of
iovec. We used to lock/lookup/unlock per fragment
which is somewhat slow if there are many fragments
to process.

With this change we lock/lookup in the `_rw_start`
call, process the incoming data in the `_rw` call
and unlock (and optionally erase) in the `_rw_end`
call.

Another way of doing this could be by passing
the array of iovec into the C++ part, but the
QEMU header which defines `struct iovec` breaks
C++.

Bug: 332748974
Test: presubmit
Change-Id: I910fab3faf9b048e7dd224b608a0e361ddf4d810
Signed-off-by: Roman Kiryanov <rkir@google.com>