Track active device plugs in ConnectorRegistry
The ConnectorRegistry now keeps track of active device plugs and
provides a method for retrieving them by name.

This allows for easier access to active devices and enables
functionality like checking if a device is currently plugged in. This
will be used by the gRPC endpoint to get access to registered devices.

Change-Id: I4300c89fc988ca35c3eefd0b5a9b7067f5b787b2