Port qemu2 command-line parsing
This ports the qemu2 command-line parsing logic to Goldfish,
allowing it to accept the same parameters as the previous
emulator.

Most of the parameters are currently unused, and this change
does not modify the existing C-style code base apart from
updating some log statements.

This change also adds unit tests for the new command-line
parsing library.

Bug: 379303514
Change-Id: I7dd5cee9459a386c198b15530bc102757c5c5ff3