docs: Add build docs for Windows, Linux, and macOS
This commit introduces documentation for building the Android Emulator on
different operating systems (Windows, Linux, and macOS).

The main goal of this commit is to improve the developer experience by
providing clear and comprehensive documentation for building the Android
Emulator on different platforms. This makes it easier for new
contributors to get started and reduces the initial setup hurdles.

Change-Id: I2514a1a0b5e4db1e84168531a66a4256837ad803