feat: Add crashpad reporting framework
This commit introduces the crashpad reporting framework to the emulator.

The crashpad framework allows for more robust and detailed crash
reporting, which will help us to better understand and address issues
users are experiencing.

This includes:

* **CrashReporter:** A singleton class that provides an interface for
  attaching data to crash reports and generating crash dumps.
* **HangDetector:** A class that monitors the emulator for hangs and
  triggers a crash report if a hang is detected.
* **CrashConsent:** An interface for managing user consent for crash
  reporting.
* **CrashUploader:** A class that handles uploading crash reports to
  a remote server.
* **Annotation tools:** A set of classes for adding annotations and
  breadcrumbs to crash reports.

These classes and interfaces will allow us to collect more detailed and
informative crash reports, which will help us to improve the stability
and reliability of the emulator.

Bug: 331193952
Change-Id: Ia4983c5ccfcc76606014d8bcfedc106fe1ba233a