feat: Add gtest_main target for Bazel file resolution
This commit introduces a `gtest_main` target specifically designed for
Google Tests that require dependencies on Bazel file resolution. This
new target should be used instead of the standard `gtest_main` when
tests need access to files specified in the 'data' section of the BUILD
file.  This enables tests to correctly locate and utilize necessary data
files within the Bazel environment.

Change-Id: I51f11370ba6c559f684657f2efc4c0f028f1e9d7