fix: Ensure child process uses same crash reporter
The child process (qemu) might not properly detect the Bazel runfiles directory,
which can cause issues when installing a crash handler.

This change ensures the child process always uses the same crash handler as the parent.

This change also forces the usage of ABSL_LOG macros, as crashpad defines the
same macros that direct to a different log engine.

Bug: 287675870
Change-Id: I99f1de222cc01451a733f822f13215194b05914c