fix: Respond with KO to all QemuMiscPipe events
The guest expects a response to all messages sent to the QemuMiscPipe,
even if the host can't interpret the message.  A lack of response causes
the guest to assume the pipe is broken.

This issue prevented the guest from signaling boot completion, which in
turn broke pytest's ability to verify emulator readiness.  Sending "KO"
for unhandled messages resolves this.

Bug: 394713760
Test: bazel run //hardware/generic/goldfish/emulator/launcher:goldfish \
     -- -avd V -vmodule "Guest*=1"  -verbose shows bootcompleted
Change-Id: Ie59d74e44870d3e6a3723fa223886a2b4482b2a7