fix: Enable C compilation for AbseilLogBridge
This commit ensures the vlog component compiles for both C and C++.  The
previous implementation had a static variable within the `_vlog_site`
function, which is not supported in C.  This change defines the static
variable at file scope, resolving the compilation issue.

Includes a set of tests to validate the expected behavior in .c files.

Change-Id: I489da3fe02a363104a39f446d1d5dfbadf505d7d