Strip released binaries and copy archive layout.
This adds stripping to all the binaries. A "remapped_pkg_files" rule is
added to "copy" a file layout created with "pkg_files" rules, but apply
it to the "transformed" version of the original files. Currently this
rule is supported by both the "stripped_binaries" rule and the
"native_symbols" rule.

With this change, you can extract a "release" archive and a
"native_symbols" archive to the same directory and get side-by-side
binaries and debug symbols.

Bug: 397510455
Change-Id: I742701ffc6aef31f06f45731e4644a85912d180d