test(emulator): Add smoke test for kernel boot
Adds a new end-to-end test that launches the emulator and verifies that
the Linux kernel starts successfully.

The test uses a minimal `minigbm` system image and runs the emulator
with the `-show-kernel` flag. It then monitors the stdout for the "Linux
version" string, which confirms that the kernel has initialized.

This provides a basic smoke test to quickly catch regressions in the
core engine boot process. The test is configured to run on Windows, Linux and macOS for both x86_64 and ARM64 ABIs.

Bug: 416062361
Change-Id: I087b01ef720154c0dc66311f4f223805acdf6b9f