test(async): Make LibraryTest more robust
The previous implementation of the
`WaitForEmptyLibraryTestWaitsUntilFinished` test was prone to flakiness
due to its reliance on `std::this_thread::sleep_for` and
a `std::condition_variable`.

This change refactors the test to use absl::Notification. This allows
for more deterministic synchronization between the worker thread and the
main test thread, eliminating the race condition and making the test
more stable.

Change-Id: I3274838d2fef9659846e7fcfcf61a053c9e720ff