build(breakpad): Use --f flag to specify dump_syms output
The dump_syms tool supports writing the output directly to a file
using the `--f` flag, which is more robust than redirecting stdout.

This change updates the `breakpad_symbols` rule to use this flag
instead of shell redirection. This avoids potential issues with
buffering and error handling that can occur with stdout redirection.

Bug: 422829578
Change-Id: I726e44b5817980abb2e20d17414411a379c19a59