fix: Account for program directory when deriving bundled executables
Previously, the system only considered the launcher directory when
searching for bundled executables. This could lead to issues where the
launcher directory differs from the actual program directory. This
change updates the `findBundledExecutable` function to also consider the
program directory, ensuring that bundled executables can be found
correctly regardless of where the launcher is located. This makes the
search more robust.

Bug: 432033514
Change-Id: I80f27a9bc1f47f8a0f070e50e90542f7f89e79dd