disk: remove iothread
Turning off iothread, as it appears to work only on Linux,
and is broken on Mac, as sometimes, it shows extremely high
IOW and blocks even adb as well (as collateral damage as
system server cannot make progress)

Test after this cl:

on Linux, there is not much changes in terms of boot speed
on Mac, this does slow down, sometimes, from 8sec to 20+ sec;
however, it is more stable.

TODO: we may want to revisit to make mac boot faster,
but not urgent.

Bug: 439608639

Change-Id: Id04f23d793919c6074617e3e45acd0b1f36ecbd2