feat(async): Add EventLoopDispatcher for event-based thread switching
This commit introduces `EventLoopDispatcher`, a new class that enables the
event-based transition of execution from one thread to another. This is
particularly useful for situations where a callback needs to be executed on a
specific event loop thread, rather than the thread that fired the event.

This will allow us to marshall events from a qemu thread to a libuv
thread and vice versa.

Bug: 438807361
Change-Id: Ib0eddcd3c9073a27f4a824d3d47ce22799a16a5a