fix(async): Ensure FIFO task execution in QemuEventLoop
QEMU's bottom-half (BH) scheduler operates in a Last-In, First-Out
(LIFO) manner. This violates the `EventLoop` contract, which requires
that tasks posted sequentially are executed in a First-In, First-Out
(FIFO) order.

Previously, each call to `post()` would schedule a new, separate BH,
leading to tasks potentially executing in the reverse order of their
submission.

This change introduces a message pump to enforce FIFO execution. All
tasks submitted via `post()` are now added to an internal queue. A
single, shared BH is scheduled to drain this queue, guaranteeing that
tasks are executed in the order they were posted.

Adds an addition test to make sure we can schedule tasks from within a
task callback.

Bug: 438807361
Change-Id: Id62cea481a78f7f56a9bae866784bd9f2a762530