update gles and vulkan libs for linux and macos
There are some duplication of icd and vk libs,
which seems necessary to get things working properly.

will need to clean up later on.

-gpu host will use host gl libs
default is angle (on top of swiftshader)

also, need to add windows equivalent, but
for now, there is no way to verify it, so
will need to add it later after windows boots.

Change-Id: Iec066f0f6b901d46ac42b7336c2566961ce1c144