Resolve most path lookups early during launcher start-up
This way we can fail earlier when necessary files are missing.
Also move path lookups into a single location to make it easier to
reason about and reduce some duplication of effort.

Also fix the launcher "data" layout so that the relative locations of
files when running under bazel more closely matches that of the release
zip. This means we can reduce the number of "inBazel" special cases.

Bug: 436554408
Bug: 432262765
Change-Id: I9e91aa68aed1241440860a755c8daf5d96cd65c8