refactor(event_loop): Remove sleep_for and timing-based logic from tests
This change refactors `EventLoopTest.RescheduleRepeatingTimer` to
replace `std::this_thread::sleep_for` and `absl::Now` with
`std::promise` and `std::future`.

The use of `std::this_thread::sleep_for` in tests introduces flakiness,
as it depends on the system scheduler and is not guaranteed to wake up
at the precise time requested. This can cause tests to fail
intermittently due to timing variations.

This is especially problematic on our build bots that are under heavy
load.

Change-Id: Ife63e69b7c6edc6acf0ae6b38eae35623be8480c