feat: Run local boot tests in exclusive mode
We have anecdotal evidence that boot tests are running close to their
timeouts, which leads to incorrect measurements and occasional test
failures.

This change introduces the `exclusive-if-local` tag to force the test to
run in exclusive mode if executed locally, ensuring that no other
actions are being performed on the machine. This avoids resource
contention, leading to more stable and accurate test results.

Change-Id: I39e08f738a35bb20cf60b973e9cb027c06a90a6d